Add point light, fix plane issue

+++ src/light.cpp
@@ -12,16 +12,16 @@ Light::Light(const Vector3 &t_origin) : m_origin(t_origin)
 {
 }
 
-Light::Light(const Vector3 &t_origin, unsigned int t_intensity) : m_origin{t_origin}, m_intensity{t_intensity}
+Light::Light(const Vector3 &t_origin, double t_intensity) : m_origin{t_origin}, m_intensity{t_intensity}
 {
 }
 
-void Light::SetIntensity(unsigned int t_intensity)
+void Light::SetIntensity(double t_intensity)
 {
     m_intensity = t_intensity;
 }
 
-unsigned int Light::GetIntensity() const
+double Light::GetIntensity() const
 {
     return m_intensity;
 }
